vigie-immo-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-Limiter==3.5.0
requests==2.31.0
geopy==2.4.1
gunicorn==21.2.0
psycopg2-binary==2.9.11
pyproj==3.7.1
PyJWT==2.8.0
bcrypt==4.1.2
